+++ Runtime/BoolVariable.cs
@@ -6,5 +6,8 @@ namespace UnderLogic.Variables
     public class BoolVariable : RuntimeVariable<bool>
     {
         public void Toggle() => SetValue(!Value);
+
+        public void CopyFrom(BoolVariable other) => Value = other.Value;
+        public void CopyTo(BoolVariable other) => other.Value = Value;
     }
 }