+++ Runtime/References/DoubleReference.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace UnderLogic.Variables.References
+{
+    [Serializable]
+    public class DoubleReference : RuntimeReference<double, DoubleVariable> { }
+}