+++ Runtime/References/TransformReference.cs
@@ -1,0 +1,8 @@
+using System;
+using UnityEngine;
+
+namespace UnderLogic.Variables.References
+{
+    [Serializable]
+    public class TransformReference : RuntimeReference<Transform, TransformVariable> { }
+}