+++ Runtime/StringVariable.cs
@@ -1,3 +1,4 @@
+using System.Globalization;
 using UnityEngine;
 
 namespace UnderLogic.Variables
@@ -7,5 +8,13 @@ namespace UnderLogic.Variables
     {
         public bool IsNullOrEmpty => string.IsNullOrEmpty(Value);
         public bool IsNullOrWhiteSpace => string.IsNullOrWhiteSpace(Value);
+
+        public void CopyFrom(StringVariable other) => Value = other.Value;
+        public void CopyTo(StringVariable other) => other.Value = Value;
+        
+        public void Stringify(bool value) => Value = value.ToString();
+        public void Stringify(int value) => Value = value.ToString();
+        public void Stringify(float value) => Value = value.ToString(CultureInfo.CurrentCulture);
+        public void Stringify(double value) => Value = value.ToString(CultureInfo.CurrentCulture);
     }
 }